Initial Work
Still figuring github out, haven't really used it for any projects in the past, but hoping to learn to use it with this project.
For now, mostly still playing with c, and now html, but hope to expand this. Tomorrow, intend to comment code pretty thoroughly. This was built from scratch in the past six hours, which have all been pretty late at night, but it works, so hopefully when I'm awake and more lucid I can better explain how it works.